Facade pattern fix
Supporting functions were moved to Facade class from MainWindow. Now client-class MainWindow only has updateAllWidget func.

+++ lab02/SmartHouseSystem/FloorHeating.h
@@ -29,7 +29,7 @@ public:
 
 private:
     std::string location;
-    double temperature;
+    int temperature;
     bool isOn;
 };
